Assmessment taken for Tata Google India Udacity Android Nanodegree Program
Instructions

Imagine we have a class called Human that represents a human.

This class holds some information about a person, such as his or her
name and age. We also have a way for a person to introduce themselves,
sharing their name and age.

What we need you to do: Create a class, Superhero, as a subclass of
Human.

The Superhero class should contain a new string that stores the
alter-ego name for these superheroic humans.

Next, you’ll modify the existing constructor to take in this alterEgo as
an additional parameter.

Then, you’ll override the introduce method to return the alterEgo in
addition to the name and age.

Finally. you’ll implement the Comparable interface to sort instances of
Human by age.